+++ packages/http-framework/src/lib/errors/index.ts
@@ -1,0 +1,5 @@
+export * from './ArgumentError.js';
+export * from './ChatInputRouterError.js';
+export * from './Identifiers.js';
+export * from './PreconditionError.js';
+export * from './UserError.js';