+++ packages/http-framework-i18n/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## 1.2.6
+
+### Patch Changes
+
+- [#143](https://github.com/wolfstar-project/stars-components/pull/143) [`09dd484`](https://github.com/wolfstar-project/stars-components/commit/09dd484926963690f84c39a24a2069b2622624dc) - Deprecate `@wolfstar/http-framework-i18n` in favour of
+  [`@wolfstar/plugin-i18next`](https://www.npmjs.com/package/@wolfstar/plugin-i18next), the official
+  `@wolfstar/http-framework` plugin that supersedes it. No further feature releases are planned for this package. Thanks [@RedStar071](https://github.com/RedStar071)!
+
+## 1.2.5
+
+### Patch Changes
+
+- [#107](https://github.com/wolfstar-project/stars-components/pull/107) [`dd057a9`](https://github.com/wolfstar-project/stars-components/commit/dd057a9096cbbaa0d80a69de6b4f10a838cdfadf) - Restore npm provenance attestation on publish for all packages Thanks [@RedStar071](https://github.com/RedStar071)!
+- Updated dependencies [[`dd057a9`](https://github.com/wolfstar-project/stars-components/commit/dd057a9096cbbaa0d80a69de6b4f10a838cdfadf)]:
+    - @wolfstar/i18next-backend@2.0.10
+
+## 1.2.4
+
+### Patch Changes
+
+- [#93](https://github.com/wolfstar-project/stars-components/pull/93) [`adce4cb`](https://github.com/wolfstar-project/stars-components/commit/adce4cb983e7f60d23bbd6d13f66adba4e2a08f5) - chore: upgrade tsdown to 0.22.14 and migrate `deps.skipNodeModulesBundle` to `deps.neverBundle` Thanks [@RedStar071](https://github.com/RedStar071)!
+- Updated dependencies [[`adce4cb`](https://github.com/wolfstar-project/stars-components/commit/adce4cb983e7f60d23bbd6d13f66adba4e2a08f5)]:
+    - @wolfstar/i18next-backend@2.0.9
+
 ## 1.2.3
 
 ### Patch Changes