+++ packages/shared-http-pieces/CHANGELOG.md
@@ -1,5 +1,34 @@
 # Changelog
 
+## 1.2.9
+
+### Patch Changes
+
+- [#112](https://github.com/wolfstar-project/stars-components/pull/112) [`1ed856a`](https://github.com/wolfstar-project/stars-components/commit/1ed856a81eb18bdc2bc47e91ffd438e939cd7591) - fix(deps): update dependency @sentry/node to v10 Thanks [@renovate](https://github.com/apps/renovate)!
+- Updated dependencies [[`abf7f77`](https://github.com/wolfstar-project/stars-components/commit/abf7f77462ae91c9840a08e273899e4027f2253a)]:
+    - @wolfstar/env-utilities@2.0.7
+    - @wolfstar/http-framework@3.1.4
+
+## 1.2.8
+
+### Patch Changes
+
+- [#107](https://github.com/wolfstar-project/stars-components/pull/107) [`dd057a9`](https://github.com/wolfstar-project/stars-components/commit/dd057a9096cbbaa0d80a69de6b4f10a838cdfadf) - Restore npm provenance attestation on publish for all packages Thanks [@RedStar071](https://github.com/RedStar071)!
+- Updated dependencies [[`dd057a9`](https://github.com/wolfstar-project/stars-components/commit/dd057a9096cbbaa0d80a69de6b4f10a838cdfadf)]:
+    - @wolfstar/env-utilities@2.0.5
+    - @wolfstar/http-framework@3.1.2
+    - @wolfstar/http-framework-i18n@1.2.5
+
+## 1.2.7
+
+### Patch Changes
+
+- [#96](https://github.com/wolfstar-project/stars-components/pull/96) [`fda4e1d`](https://github.com/wolfstar-project/stars-components/commit/fda4e1d22edff4c516a599cec45294b1dfbde40b) - chore: migrate localization tooling from Crowdin to Tolgee Thanks [@RedStar071](https://github.com/RedStar071)!
+- Updated dependencies [[`adce4cb`](https://github.com/wolfstar-project/stars-components/commit/adce4cb983e7f60d23bbd6d13f66adba4e2a08f5)]:
+    - @wolfstar/env-utilities@2.0.4
+    - @wolfstar/http-framework-i18n@1.2.4
+    - @wolfstar/http-framework@3.1.1
+
 ## 1.2.6
 
 ### Patch Changes